pyproject.toml
[project]
name = "ring-artifact-unet"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "lightning>=2.6.0",
    "matplotlib>=3.10.8",
    "notebook>=7.5.2",
    "tensorboard>=2.20.0",
    "tifffile>=2026.1.14",
    "torch>=2.9.1",
    "torchmetrics>=1.8.2",
    "torchvision>=0.24.1",
]

[project.optional-dependencies]
synthesize = [
    "astra-toolbox>=2.4.1",
]


[tool.uv.sources]
torch = [
  { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
  { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[[tool.uv.index]]
name = "aliyun-pypi"
url = "https://mirrors.aliyun.com/pypi/simple/"
explicit = false

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
]
